==> 02_3_falhas_brutas_15min.py <==
import json
import pandas as pd
import os

#detalha horario de erros grosseiros
#revisao ate 30/11/2020
#pd.date_range(start="2019-01-11 08:00:00",end="2019-01-15 11:00:00",
#              freq='15min').format(formatter=lambda x: x.strftime(
#                   '%Y-%m-%d %H:%M:%S'))
erros = {'Rio_Negro' : [
    pd.date_range(start="1997-10-24T03:15:00Z",end="1997-10-31T11:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1998-01-27T20:00:00Z',
    pd.date_range(start="1998-06-04T23:15:00Z",end="1998-06-05T02:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1998-06-05T13:15:00Z",end="1998-06-05T16:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1998-06-10T04:00:00Z",end="1998-06-12T18:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1998-07-03T02:30:00Z',
    '1998-07-03T15:15:00Z',
    pd.date_range(start="1998-09-14T18:30:00Z",end="1998-09-15T18:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1998-12-31T02:45:00Z',
    pd.date_range(start="1999-06-01T17:00:00Z",end="1999-06-01T18:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1999-09-03T20:00:00Z',
    pd.date_range(start="1999-10-09T12:15:00Z",end="1999-10-10T18:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1999-10-13T16:15:00Z',
    '1999-10-13T17:15:00Z',
    '1999-12-06T17:15:00Z',
    '1999-12-06T17:30:00Z',
    pd.date_range(start="2000-06-05T17:15:00Z",end="2000-06-05T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-05-03T23:45:00Z",end="2001-05-04T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2001-05-09T14:45:00Z',
    pd.date_range(start="2001-05-14T17:30:00Z",end="2001-05-14T18:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2002-10-31T17:45:00Z',
    pd.date_range(start="2003-04-30T17:00:00Z",end="2003-05-07T10:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2003-05-10T01:30:00Z',
    '2003-07-03T08:00:00Z',
    '2003-07-24T03:30:00Z',
    '2003-07-25T01:15:00Z',
    '2003-08-05T04:30:00Z',
    '2003-09-13T22:45:00Z',
    '2003-10-07T14:00:00Z',
    '2004-07-31T10:30:00Z',
    '2004-08-24T19:15:00Z',
    '2004-10-01T14:30:00Z',
    '2005-01-10T10:45:00Z',
    '2005-03-24T06:30:00Z',
    '2005-06-17T18:15:00Z',
    '2005-06-27T13:00:00Z',
    '2005-06-29T14:45:00Z',
    '2005-07-01T14:00:00Z',
    '2005-07-14T06:30:00Z',
    '2005-08-07T22:15:00Z',
    '2005-08-29T17:45:00Z',
    '2005-09-03T11:30:00Z',
    '2005-09-07T03:30:00Z',
    '2005-09-08T22:30:00Z',
    pd.date_range(start="2005-11-11T15:00:00Z",end="2005-11-11T17:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2005-11-28T17:15:00Z",end="2005-11-28T18:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2006-04-22T21:45:00Z',
    '2006-04-26T04:45:00Z',
    pd.date_range(start="2006-06-23T13:30:00Z",end="2006-06-28T10:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2007-08-10T17:15:00Z',
    '2007-11-28T18:30:00Z',
    '2007-12-15T18:15:00Z',
    '2007-12-27T14:30:00Z',
    '2008-02-22T02:15:00Z',
    '2008-03-31T05:15:00Z',
    '2008-04-16T16:30:00Z',
    '2008-05-21T09:00:00Z',
    '2008-06-07T10:45:00Z',
    '2008-08-27T14:30:00Z',
    '2008-08-30T02:30:00Z',
    '2009-03-18T09:15:00Z',
    '2009-03-28T22:15:00Z',
    '2009-10-02T19:30:00Z',
    '2009-11-06T13:15:00Z',
    '2009-12-04T14:15:00Z',
    '2010-04-08T05:15:00Z',
    '2010-04-29T19:45:00Z',
    pd.date_range(start="2010-04-30T16:45:00Z",end="2010-04-30T17:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2010-05-05T05:30:00Z",end="2010-05-06T16:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2011-05-06T14:45:00Z',
    '2012-02-23T15:30:00Z',
    '2012-04-16T12:00:00Z',
    '2012-04-16T12:30:00Z',
    '2012-04-21T12:30:00Z',
    '2012-04-25T03:45:00Z',
    '2012-06-06T15:00:00Z',
    '2012-07-10T21:00:00Z',
    '2013-05-22T14:45:00Z',
    '2013-08-30T15:00:00Z',
    '2013-11-19T18:30:00Z',
    pd.date_range(start="2014-01-16T13:15:00Z",end="2014-01-17T16:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2014-06-11T20:30:00Z',
    pd.date_range(start="2017-05-14T03:15:00Z",end="2017-05-14T04:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-06-01T15:30:00Z',
    pd.date_range(start="2017-06-02T00:15:00Z",end="2017-06-02T15:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2020-09-22T14:15:00Z",end="2020-09-22T18:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
                        ],

         'Porto_Amazonas' : [
    '1997-11-20T19:30:00Z',
    pd.date_range(start="1998-02-13T16:00:00Z",end="1998-02-13T18:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1999-07-14T15:00:00Z",end="1999-07-15T16:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2000-01-22T02:15:00Z',
    pd.date_range(start="2001-03-02T15:45:00Z",end="2001-03-06T18:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2001-07-19T17:30:00Z',
    '2001-08-07T02:15:00Z',
    '2001-08-07T03:45:00Z',
    '2001-08-07T09:30:00Z',
    '2001-08-07T14:00:00Z',
    '2001-08-07T14:15:00Z',
    '2001-08-08T01:15:00Z',
    '2001-08-08T02:45:00Z',
    pd.date_range(start="2001-08-08T12:15:00Z",end="2001-08-08T12:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2001-08-08T14:15:00Z',
    '2001-08-08T17:00:00Z',
    pd.date_range(start="2001-08-09T13:00:00Z",end="2001-08-09T14:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2002-02-03T02:30:00Z',
    '2002-02-12T16:45:00Z',
    '2002-02-12T17:00:00Z',
    '2002-03-24T05:00:00Z',
    pd.date_range(start="2002-06-15T16:45:00Z",end="2002-06-15T18:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2003-06-11T17:30:00Z',
    pd.date_range(start="2003-09-04T15:15:00Z",end="2003-09-04T18:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2004-05-13T20:45:00Z',
    '2004-08-31T09:45:00Z',
    '2004-11-30T06:45:00Z',
    '2004-11-30T06:45:00Z',
    '2005-04-15T21:45:00Z',
    '2005-06-22T09:45:00Z',
    '2005-07-18T06:00:00Z',
    '2005-07-25T22:45:00Z',
    '2005-12-04T17:30:00Z',
    '2006-01-15T13:45:00Z',
    '2007-08-14T02:30:00Z',
    '2008-09-29T10:15:00Z',
    '2009-03-27T20:15:00Z',
    '2009-04-24T06:15:00Z',
    '2009-10-09T13:45:00Z',
    pd.date_range(start="2010-02-16T21:45:00Z",end="2010-02-25T17:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2010-05-13T17:45:00Z',
    '2010-06-26T21:15:00Z',
    '2010-06-27T02:15:00Z',
    '2010-06-29T11:15:00Z',
    '2010-07-02T18:00:00Z',
    '2010-08-02T03:30:00Z',
    '2010-08-05T08:00:00Z',
    '2010-08-16T07:15:00Z',
    pd.date_range(start="2010-09-03T14:30:00Z",end="2010-09-09T18:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2010-09-10T19:15:00Z',
    '2010-09-10T20:15:00Z',
    '2010-09-15T11:15:00Z',
    '2010-09-15T12:15:00Z',
    '2010-09-20T11:30:00Z',
    '2010-09-20T12:30:00Z',
    '2010-09-24T13:00:00Z',
    '2010-09-25T13:15:00Z',
    '2010-09-26T17:15:00Z',
    '2010-10-05T11:45:00Z',
    '2010-10-09T02:45:00Z',
    '2010-10-30T04:15:00Z',
    '2010-11-02T23:15:00Z',
    '2010-11-05T22:00:00Z',
    '2010-11-12T20:30:00Z',
    '2010-11-14T04:30:00Z',
    '2010-11-14T05:30:00Z',
    '2010-11-22T19:45:00Z',
    '2010-11-22T20:00:00Z',
    '2010-11-29T09:15:00Z',
    '2010-12-01T19:15:00Z',
    '2010-12-01T21:15:00Z',
    '2010-12-02T01:15:00Z',
    '2010-12-05T21:30:00Z',
    '2010-12-08T10:00:00Z',
    '2010-12-15T20:30:00Z',
    '2010-12-24T06:15:00Z',
    '2011-01-05T00:15:00Z',
    '2011-01-08T14:30:00Z',
    '2011-01-08T17:30:00Z',
    '2011-01-11T07:45:00Z',
    '2011-01-13T23:15:00Z',
    '2011-01-20T00:30:00Z',
    '2011-01-21T01:30:00Z',
    '2011-01-23T07:15:00Z',
    '2011-02-11T05:00:00Z',
    '2011-02-11T12:00:00Z',
    '2011-02-11T15:00:00Z',
    '2011-02-25T13:15:00Z',
    '2011-03-01T03:15:00Z',
    '2011-03-03T10:15:00Z',
    '2011-03-06T10:15:00Z',
    '2011-03-08T23:30:00Z',
    '2011-03-24T09:00:00Z',
    '2011-03-30T11:30:00Z',
    '2011-04-13T02:30:00Z',
    '2011-04-20T03:30:00Z',
    '2011-04-30T16:00:00Z',
    '2011-06-09T07:00:00Z',
    '2011-06-13T16:00:00Z',
    '2011-06-14T14:00:00Z',
    '2011-06-19T11:15:00Z',
    '2011-06-28T14:30:00Z',
    '2011-06-30T17:30:00Z',
    '2011-07-01T15:30:00Z',
    '2011-07-06T06:30:00Z',
    '2011-07-11T06:45:00Z',
    '2011-07-11T13:45:00Z',
    '2011-07-18T09:00:00Z',
    '2011-07-18T10:00:00Z',
    '2011-07-24T05:30:00Z',
    '2011-07-28T14:15:00Z',
    '2011-07-28T23:15:00Z',
    '2011-08-16T23:15:00Z',
    '2011-08-17T00:15:00Z',
    '2011-08-19T04:00:00Z',
    '2011-08-28T14:15:00Z',
    '2011-08-29T01:15:00Z',
    '2011-09-01T17:30:00Z',
    '2011-09-04T18:30:00Z',
    '2011-09-15T00:30:00Z',
    '2011-09-18T11:45:00Z',
    '2011-09-18T16:45:00Z',
    '2011-09-20T20:00:00Z',
    '2011-09-20T21:00:00Z',
    '2011-10-06T11:15:00Z',
    '2011-10-08T11:30:00Z',
    '2011-10-09T11:30:00Z',
    '2011-10-16T19:00:00Z',
    '2011-10-21T07:30:00Z',
    '2011-10-23T04:15:00Z',
    '2011-11-01T18:00:00Z',
    '2011-11-11T11:15:00Z',
    '2011-11-13T03:15:00Z',
    '2011-11-14T06:15:00Z',
    '2011-11-21T21:30:00Z',
    '2011-12-09T00:15:00Z',
    '2011-12-16T03:00:00Z',
    '2011-12-31T01:30:00Z',
    pd.date_range(start="2012-01-05T17:45:00Z",end="2012-01-06T01:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2012-01-11T13:45:00Z",end="2012-01-11T14:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2012-01-21T01:15:00Z',
    '2012-01-23T07:15:00Z',
    '2012-01-28T21:15:00Z',
    '2012-01-30T15:15:00Z',
    '2012-02-05T07:30:00Z',
    '2012-02-06T10:30:00Z',
    '2012-02-15T10:15:00Z',
    '2012-03-16T08:15:00Z',
    '2012-03-22T10:30:00Z',
    '2012-03-25T20:00:00Z',
    '2012-04-07T18:15:00Z',
    pd.date_range(start="2012-04-08T17:00:00Z",end="2012-04-09T17:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2012-10-23T20:15:00Z',
    pd.date_range(start="2014-11-01T19:30:00Z",end="2014-11-04T17:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2015-02-18T20:30:00Z',
    '2015-02-19T02:15:00Z',
    '2016-04-18T17:45:00Z',
    pd.date_range(start="2017-11-17T01:15:00Z",end="2017-11-17T02:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2017-12-27T02:30:00Z",end="2017-12-28T12:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2019-03-22T12:45:00Z",end="2019-03-22T13:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2019-09-09T23:15:00Z",end="2019-09-10T00:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2019-09-10T17:45:00Z",end="2019-09-11T12:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2020-05-03T17:00:00Z",end="2020-05-05T15:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2020-05-06T13:45:00Z",end="2020-05-06T14:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2020-12-02T17:15:00Z',
                             ],

         'Sao_Bento' : [
    '1998-02-22T01:00:00Z',
    pd.date_range(start="2001-05-13T00:30:00Z",end="2001-05-14T07:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-05-14T19:45:00Z",end="2001-05-14T20:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-05-14T23:45:00Z",end="2001-05-15T02:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-05-16T20:00:00Z",end="2001-05-16T23:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-06-20T19:15:00Z",end="2001-06-24T09:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-06-29T16:15:00Z",end="2001-06-30T15:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2002-06-08T14:00:00Z',
    '2002-09-26T14:15:00Z',
    '2003-09-10T17:15:00Z',
    '2003-09-20T22:00:00Z',
    '2003-10-10T02:45:00Z',
    '2003-11-16T00:15:00Z',
    '2004-06-14T01:15:00Z',
    '2004-06-30T09:15:00Z',
    '2004-07-31T02:30:00Z',
    '2004-08-29T02:00:00Z',
    pd.date_range(start="2004-11-12T20:30:00Z",end="2004-11-13T12:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2004-11-14T05:45:00Z",end="2004-11-14T06:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2004-11-14T14:15:00Z",end="2004-11-16T13:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2004-11-21T01:00:00Z',
    '2004-11-21T19:00:00Z',
    '2004-11-23T23:30:00Z',
    pd.date_range(start="2004-11-27T01:15:00Z",end="2004-11-29T22:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2005-06-08T02:30:00Z',
    '2005-09-13T18:15:00Z',
    '2005-10-07T02:30:00Z',
    '2006-03-15T14:00:00Z',
    '2006-08-08T02:30:00Z',
    '2007-09-04T21:30:00Z',
    '2007-09-18T21:45:00Z',
    '2007-10-23T00:00:00Z',
    '2008-01-02T03:45:00Z',
    '2008-05-13T05:30:00Z',
    '2008-05-21T12:00:00Z',
    '2008-07-21T13:15:00Z',
    '2008-09-09T06:45:00Z',
    '2008-09-11T19:30:00Z',
    '2008-09-24T19:15:00Z',
    '2008-11-04T01:00:00Z',
    '2009-02-22T18:45:00Z',
    '2009-02-24T23:30:00Z',
    '2009-05-08T00:00:00Z',
    '2009-06-09T21:00:00Z',
    '2009-07-13T09:00:00Z',
    '2009-08-07T06:30:00Z',
    '2009-08-07T17:30:00Z',
    '2009-08-09T03:00:00Z',
    '2009-09-14T03:15:00Z',
    '2009-09-14T07:15:00Z',
    '2009-09-20T17:30:00Z',
    '2009-10-23T12:30:00Z',
    '2009-10-26T12:30:00Z',
    '2009-10-31T17:00:00Z',
    '2009-11-04T15:15:00Z',
    '2009-11-04T23:00:00Z',
    '2009-11-07T09:15:00Z',
    '2009-11-22T10:00:00Z',
    '2009-11-24T12:15:00Z',
    '2010-01-23T06:30:00Z',
    '2010-02-26T10:30:00Z',
    '2010-02-27T11:00:00Z',
    '2010-02-28T06:00:00Z',
    '2010-03-09T15:00:00Z',
    '2010-04-10T07:00:00Z',
    '2010-05-16T07:00:00Z',
    '2010-05-21T05:00:00Z',
    '2010-05-29T16:15:00Z',
    '2010-06-28T17:15:00Z',
    '2010-10-20T18:15:00Z',
    '2010-10-22T03:45:00Z',
    '2010-11-05T20:45:00Z',
    '2010-11-12T14:30:00Z',
    '2011-03-05T23:45:00Z',
    '2012-06-14T16:30:00Z',
    pd.date_range(start="2012-06-25T18:30:00Z",end="2012-06-25T19:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2012-06-28T06:15:00Z",end="2012-10-02T14:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2012-10-21T00:00:00Z',
    '2013-03-24T22:45:00Z',
    '2013-12-06T00:45:00Z',
    '2014-11-02T01:00:00Z',
    '2014-11-22T00:15:00Z',
    '2015-02-24T23:15:00Z',
    pd.date_range(start="2015-08-26T18:00:00Z",end="2015-08-27T05:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2016-08-30T17:30:00Z',
    pd.date_range(start="2017-06-22T16:45:00Z",end="2017-06-22T17:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-01-11T11:15:00Z',
    '2019-03-15T14:45:00Z',
    pd.date_range(start="2019-03-18T19:45:00Z",end="2019-03-18T23:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-03-19T09:30:00Z',
    pd.date_range(start="2019-03-19T18:30:00Z",end="2019-03-21T20:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-05-19T16:45:00Z',
    '2019-05-21T19:15:00Z',
    '2019-05-26T01:15:00Z',
    '2019-05-26T17:15:00Z',
    '2019-05-27T02:00:00Z',
    '2019-05-27T18:15:00Z',
    '2019-05-27T19:00:00Z',
    '2019-05-27T20:00:00Z',
    pd.date_range(start="2019-06-01T16:00:00Z",end="2019-06-01T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2019-06-03T18:30:00Z",end="2019-06-04T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2020-11-22T23:45:00Z',
    '2020-11-24T20:00:00Z',
                        ],

         'Pontilhao' : [
    pd.date_range(start="1997-06-01T09:00:00Z",end="1997-06-01T12:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1997-07-12T05:00:00Z",end="1997-07-12T07:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1997-07-12T21:00:00Z",end="1997-07-13T00:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1997-07-22T09:00:00Z',
    pd.date_range(start="1997-08-05T16:00:00Z",end="1997-08-05T18:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1997-09-30T15:00:00Z',
    '1998-07-05T14:15:00Z',
    '1999-03-28T03:45:00Z',
    '1999-10-16T02:45:00Z',
    '1999-11-07T03:30:00Z',
    pd.date_range(start="2000-10-02T20:45:00Z",end="2000-10-02T21:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2000-11-09T23:15:00Z',
    '2000-11-28T00:00:00Z',
    '2001-03-01T01:30:00Z',
    '2001-03-16T01:00:00Z',
    '2001-08-17T14:15:00Z',
    '2002-03-16T02:15:00Z',
    '2002-04-18T13:30:00Z',
    pd.date_range(start="2002-06-02T10:45:00Z",end="2002-06-02T20:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2002-06-28T16:45:00Z",end="2002-07-03T17:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2002-09-27T23:15:00Z',
    '2002-12-13T23:30:00Z',
    '2002-12-23T22:15:00Z',
    '2003-04-05T16:30:00Z',
    '2003-05-17T18:00:00Z',
    '2003-05-17T19:00:00Z',
    '2003-05-20T19:15:00Z',
    '2003-07-02T05:00:00Z',
    pd.date_range(start="2003-09-17T08:45:00Z",end="2003-09-19T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2003-11-01T10:45:00Z',
    '2003-11-16T16:15:00Z',
    pd.date_range(start="2004-01-05T17:15:00Z",end="2004-01-10T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2004-01-25T11:00:00Z',
    '2004-02-22T07:15:00Z',
    '2004-03-04T06:45:00Z',
    '2004-03-21T02:30:00Z',
    '2004-03-25T02:15:00Z',
    '2004-04-23T18:15:00Z',
    '2004-10-18T02:15:00Z',
    '2004-10-28T22:15:00Z',
    '2004-10-30T00:30:00Z',
    '2004-11-16T14:30:00Z',
    '2005-03-29T15:00:00Z',
    '2005-04-10T03:00:00Z',
    '2005-09-15T15:30:00Z',
    pd.date_range(start="2005-10-10T18:45:00Z",end="2005-10-10T19:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2005-10-14T03:15:00Z',
    '2006-01-07T23:15:00Z',
    '2006-04-21T10:00:00Z',
    '2006-05-19T07:30:00Z',
    '2006-05-30T17:00:00Z',
    '2006-05-30T17:15:00Z',
    pd.date_range(start="2006-06-20T17:15:00Z",end="2006-06-20T19:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2006-08-21T07:00:00Z',
    '2006-11-16T11:00:00Z',
    '2006-12-28T07:00:00Z',
    '2007-03-19T02:15:00Z',
    '2007-06-30T03:45:00Z',
    '2007-06-30T22:45:00Z',
    '2007-08-19T10:45:00Z',
    '2008-02-18T19:30:00Z',
    '2008-03-26T11:30:00Z',
    '2008-04-23T02:45:00Z',
    '2008-08-29T19:15:00Z',
    '2008-10-16T23:30:00Z',
    '2008-10-27T10:15:00Z',
    '2008-11-23T08:30:00Z',
    '2009-02-15T18:30:00Z',
    '2009-03-26T07:00:00Z',
    '2009-04-17T03:00:00Z',
    '2009-07-06T06:15:00Z',
    '2009-07-12T08:30:00Z',
    '2009-07-27T05:00:00Z',
    '2009-09-05T03:15:00Z',
    '2009-09-19T20:15:00Z',
    '2009-09-20T05:15:00Z',
    '2009-09-23T01:30:00Z',
    '2009-09-27T00:15:00Z',
    '2009-10-19T05:00:00Z',
    '2009-11-05T02:00:00Z',
    '2009-12-19T15:15:00Z',
    '2010-02-09T10:30:00Z',
    '2010-02-21T18:15:00Z',
    '2010-02-24T00:15:00Z',
    '2010-03-18T16:15:00Z',
    '2010-04-04T09:30:00Z',
    '2010-04-25T09:00:00Z',
    '2010-04-27T22:15:00Z',
    '2010-04-28T23:30:00Z',
    '2010-05-07T13:00:00Z',
    pd.date_range(start="2010-05-12T18:45:00Z",end="2010-05-13T15:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2010-06-03T11:15:00Z',
    '2010-06-12T21:00:00Z',
    '2010-06-16T20:00:00Z',
    '2010-07-12T17:30:00Z',
    '2010-08-11T08:30:00Z',
    '2010-08-17T16:00:00Z',
    '2010-10-22T14:30:00Z',
    '2010-11-07T14:15:00Z',
    '2011-01-18T14:15:00Z',
    '2011-04-07T19:30:00Z',
    '2011-04-07T19:45:00Z',
    '2011-04-08T23:15:00Z',
    '2011-04-15T02:45:00Z',
    '2011-04-23T16:15:00Z',
    '2011-04-26T00:15:00Z',
    '2011-07-08T20:00:00Z',
    '2011-07-26T05:00:00Z',
    '2011-07-26T10:00:00Z',
    '2011-07-30T07:15:00Z',
    '2011-08-14T22:30:00Z',
    '2011-08-16T05:30:00Z',
    '2011-08-16T11:00:00Z',
    '2011-08-16T23:00:00Z',
    '2011-08-19T11:00:00Z',
    '2011-08-22T18:30:00Z',
    '2011-08-24T20:00:00Z',
    '2011-08-26T03:15:00Z',
    '2011-09-06T07:00:00Z',
    '2011-09-06T18:15:00Z',
    '2011-09-15T02:45:00Z',
    '2011-09-17T06:15:00Z',
    '2011-10-13T21:00:00Z',
    pd.date_range(start="2011-10-18T14:30:00Z",end="2011-10-20T18:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2011-10-25T15:45:00Z',
    '2011-10-27T16:45:00Z',
    '2011-10-27T23:45:00Z',
    '2011-10-28T11:45:00Z',
    '2011-11-03T05:00:00Z',
    '2011-11-03T10:00:00Z',
    '2011-11-04T13:00:00Z',
    '2011-11-11T22:30:00Z',
    '2011-11-14T00:30:00Z',
    '2011-11-18T23:00:00Z',
    '2011-11-19T06:00:00Z',
    '2011-11-22T06:45:00Z',
    '2011-11-23T10:45:00Z',
    '2011-11-23T15:45:00Z',
    '2011-11-24T23:30:00Z',
    '2011-11-25T00:00:00Z',
    '2011-12-03T07:45:00Z',
    '2011-12-03T08:45:00Z',
    '2011-12-04T06:45:00Z',
    '2011-12-04T12:45:00Z',
    '2011-12-10T17:45:00Z',
    '2011-12-11T01:45:00Z',
    '2011-12-11T10:45:00Z',
    '2011-12-11T19:45:00Z',
    '2012-01-06T11:45:00Z',
    '2012-01-10T06:45:00Z',
    '2012-01-12T04:45:00Z',
    '2012-01-13T03:45:00Z',
    '2012-01-24T21:45:00Z',
    '2012-01-29T03:45:00Z',
    '2012-01-29T16:45:00Z',
    '2012-02-01T00:45:00Z',
    '2012-02-06T06:45:00Z',
    '2012-02-17T17:45:00Z',
    '2012-02-17T20:45:00Z',
    '2012-02-19T06:00:00Z',
    '2012-03-12T16:15:00Z',
    '2012-03-12T17:00:00Z',
    '2012-04-29T03:00:00Z',
    '2012-04-29T10:15:00Z',
    '2012-04-30T04:45:00Z',
    '2012-05-01T01:15:00Z',
    '2012-05-01T02:45:00Z',
    pd.date_range(start="2012-05-01T05:30:00Z",end="2012-05-01T06:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2012-05-01T08:45:00Z',
    '2012-05-02T15:15:00Z',
    '2015-02-06T01:15:00Z',
    pd.date_range(start="2016-02-11T16:15:00Z",end="2016-02-11T16:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2016-04-05T09:00:00Z',
    '2016-04-23T21:00:00Z',
    pd.date_range(start="2017-04-25T18:00:00Z",end="2017-04-25T19:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2017-07-28T11:45:00Z",end="2017-07-31T12:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-12-13T18:00:00Z',
    '2017-12-13T18:15:00Z',
    pd.date_range(start="2020-11-03T18:15:00Z",end="2020-11-03T18:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
                        ],

         'Santa_Cruz_Timbo' : [
    '2006-10-18T06:15:00Z',
    '2006-10-24T10:15:00Z',
    '2006-10-31T12:00:00Z',
    '2006-11-10T23:45:00Z',
    pd.date_range(start="2006-11-11T02:00:00Z",end="2006-11-11T02:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2006-11-15T21:30:00Z',
    '2007-01-27T06:45:00Z',
    pd.date_range(start="2007-03-12T01:00:00Z",end="2007-03-12T01:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2007-04-13T10:45:00Z",end="2007-04-13T11:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2007-08-03T11:00:00Z',
    '2008-01-16T15:45:00Z',
    pd.date_range(start="2008-01-29T19:15:00Z",end="2008-01-29T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-03-13T03:00:00Z',
    '2008-03-18T17:45:00Z',
    pd.date_range(start="2008-03-18T18:15:00Z",end="2008-03-18T19:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-03-18T20:45:00Z',
    pd.date_range(start="2008-03-21T15:15:00Z",end="2008-03-21T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2008-03-21T16:30:00Z",end="2008-03-21T21:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-03-26T17:45:00Z',
    pd.date_range(start="2008-03-26T18:30:00Z",end="2008-03-26T19:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-03-26T20:30:00Z',
    '2008-03-26T21:00:00Z',
    pd.date_range(start="2008-03-27T15:00:00Z",end="2008-03-27T20:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-03-28T11:45:00Z',
    '2008-12-17T08:00:00Z',
    '2009-06-05T10:30:00Z',
    '2009-06-24T18:00:00Z',
    '2009-08-06T14:00:00Z',
    '2009-10-31T15:30:00Z',
    '2009-11-02T06:30:00Z',
    '2010-03-10T19:45:00Z',
    '2010-04-02T03:45:00Z',
    '2010-05-19T13:15:00Z',
    pd.date_range(start="2010-07-09T13:00:00Z",end="2010-07-09T17:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2010-08-05T01:30:00Z',
    '2010-08-11T13:00:00Z',
    '2010-09-25T12:30:00Z',
    '2011-02-04T01:45:00Z',
    '2011-02-10T22:00:00Z',
    '2011-02-12T02:00:00Z',
    '2011-02-22T13:45:00Z',
    '2011-02-28T05:45:00Z',
    '2011-03-01T14:45:00Z',
    '2011-03-01T15:45:00Z',
    '2011-03-04T19:45:00Z',
    '2011-03-05T20:00:00Z',
    '2011-03-09T18:00:00Z',
    '2011-03-11T02:00:00Z',
    '2011-03-14T02:00:00Z',
    '2011-03-15T21:45:00Z',
    '2011-03-21T20:45:00Z',
    '2011-03-22T05:45:00Z',
    '2011-03-24T13:45:00Z',
    '2011-04-03T08:45:00Z',
    '2011-04-03T17:45:00Z',
    '2011-04-06T05:45:00Z',
    '2011-04-07T16:45:00Z',
    '2011-04-08T18:45:00Z',
    '2011-04-09T07:45:00Z',
    '2011-04-10T14:45:00Z',
    '2011-04-10T15:45:00Z',
    '2011-04-11T04:45:00Z',
    '2011-04-16T10:00:00Z',
    '2011-04-16T15:00:00Z',
    '2011-04-17T03:00:00Z',
    '2011-04-20T23:00:00Z',
    '2011-04-24T02:00:00Z',
    '2011-04-24T06:00:00Z',
    '2011-04-24T19:00:00Z',
    '2011-04-30T15:00:00Z',
    '2011-05-01T07:00:00Z',
    '2011-05-05T13:00:00Z',
    '2011-05-10T19:45:00Z',
    '2011-05-10T21:45:00Z',
    '2011-05-11T11:45:00Z',
    '2011-05-12T02:45:00Z',
    '2011-05-12T06:45:00Z',
    '2011-05-12T13:45:00Z',
    '2011-05-12T18:45:00Z',
    '2011-05-12T20:45:00Z',
    '2011-05-14T15:45:00Z',
    '2011-05-15T10:45:00Z',
    '2011-05-20T06:45:00Z',
    '2011-05-21T04:45:00Z',
    '2011-05-26T13:45:00Z',
    '2011-05-27T15:45:00Z',
    '2011-05-27T17:45:00Z',
    '2011-05-27T21:45:00Z',
    '2011-06-14T04:00:00Z',
    '2011-06-15T12:00:00Z',
    '2011-06-18T02:00:00Z',
    '2011-06-18T16:00:00Z',
    '2011-06-19T00:00:00Z',
    '2011-06-19T08:00:00Z',
    '2011-06-22T10:45:00Z',
    '2011-06-28T21:45:00Z',
    '2011-07-01T20:45:00Z',
    '2011-07-10T07:45:00Z',
    '2011-07-10T09:45:00Z',
    '2011-07-11T11:45:00Z',
    '2011-07-11T14:45:00Z',
    '2011-07-14T03:45:00Z',
    '2011-07-16T16:00:00Z',
    '2011-07-16T17:00:00Z',
    '2011-07-25T04:45:00Z',
    '2011-07-25T06:45:00Z',
    '2011-07-27T11:45:00Z',
    '2011-08-01T02:45:00Z',
    '2011-08-01T03:45:00Z',
    '2011-08-02T15:45:00Z',
    '2011-08-04T05:45:00Z',
    '2011-08-04T10:45:00Z',
    '2011-08-04T13:45:00Z',
    '2011-08-06T10:00:00Z',
    '2011-08-06T11:00:00Z',
    '2011-08-06T15:00:00Z',
    '2011-08-06T16:00:00Z',
    '2011-08-08T23:45:00Z',
    '2011-08-11T17:00:00Z',
    '2011-08-13T12:00:00Z',
    '2011-08-13T21:00:00Z',
    '2011-08-18T06:00:00Z',
    '2011-08-19T19:00:00Z',
    '2011-08-23T06:00:00Z',
    '2011-08-25T07:45:00Z',
    '2011-08-25T15:45:00Z',
    '2011-08-27T22:45:00Z',
    '2011-08-28T02:45:00Z',
    '2011-09-19T04:45:00Z',
    '2011-09-24T12:45:00Z',
    '2011-09-28T19:45:00Z',
    '2011-09-29T10:45:00Z',
    '2011-10-05T22:45:00Z',
    '2011-10-07T02:45:00Z',
    '2011-10-08T18:45:00Z',
    '2011-10-09T04:45:00Z',
    '2011-10-13T06:45:00Z',
    '2011-10-18T09:00:00Z',
    '2011-10-19T20:00:00Z',
    '2011-10-20T02:00:00Z',
    '2011-10-20T03:00:00Z',
    '2011-10-20T04:00:00Z',
    '2011-10-22T23:15:00Z',
    '2011-10-25T08:45:00Z',
    '2011-10-25T15:45:00Z',
    '2011-10-25T16:45:00Z',
    pd.date_range(start="2012-02-05T14:45:00Z",end="2012-02-06T19:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2013-06-19T04:45:00Z",end="2013-06-19T05:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2014-06-17T15:30:00Z',
    '2014-07-17T21:00:00Z',
    '2018-11-15T19:30:00Z',
    '2018-11-25T00:30:00Z',
    '2018-12-05T19:30:00Z',
    pd.date_range(start="2019-01-29T17:30:00Z",end="2019-01-29T17:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-02-01T01:30:00Z',
    '2019-02-20T22:30:00Z',
    '2019-02-27T00:45:00Z',
    '2019-03-07T00:45:00Z',
    '2019-03-27T00:45:00Z',
    '2019-03-31T19:30:00Z',
    '2019-04-02T21:45:00Z',
    '2019-05-18T23:30:00Z',
    '2019-05-25T08:15:00Z',
    '2019-06-27T07:30:00Z',
    '2019-08-02T23:30:00Z',
    '2019-08-03T22:30:00Z',
    '2019-10-24T08:45:00Z',
    '2019-11-21T11:30:00Z',
    '2019-11-30T23:30:00Z',
    '2019-12-22T16:30:00Z',
    '2020-01-03T18:30:00Z',
    pd.date_range(start="2020-02-02T20:15:00Z",end="2020-02-02T21:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2020-02-03T02:45:00Z',
    '2020-03-04T10:30:00Z',
    '2020-03-14T09:45:00Z',
    '2020-03-20T21:45:00Z',
    '2020-03-21T11:30:00Z',
    '2020-04-25T08:30:00Z',
    '2020-04-26T21:30:00Z',
    '2020-05-02T06:45:00Z',
    '2020-05-03T15:30:00Z',
    '2020-05-11T16:30:00Z',
    '2020-05-12T20:30:00Z',
    '2020-05-13T07:30:00Z',
    '2020-07-15T16:30:00Z',
    '2020-09-13T07:30:00Z',
    '2020-09-15T14:30:00Z',
    '2020-11-30T21:30:00Z',
    pd.date_range(start="2020-12-25T10:30:00Z",end="2020-12-25T10:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
                               ],

         'Sao_Mateus_Sul' : [
    '1997-06-19T14:45:00Z',
    '1997-12-10T18:45:00Z',
    '2000-10-19T21:45:00Z',
    '2000-10-20T08:45:00Z',
    '2000-12-16T23:30:00Z',
    '2000-12-29T20:45:00Z',
    pd.date_range(start="2001-01-09T16:15:00Z",end="2001-01-11T09:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-06-01T10:30:00Z",end="2001-06-07T16:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-08-03T02:45:00Z",end="2001-08-10T21:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2002-03-07T21:15:00Z',
    pd.date_range(start="2003-05-17T18:45:00Z",end="2003-05-17T19:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2003-08-07T21:45:00Z',
    '2004-01-05T07:15:00Z',
    '2004-01-28T02:45:00Z',
    '2004-04-09T15:15:00Z',
    '2004-05-29T18:00:00Z',
    '2004-09-02T11:30:00Z',
    '2004-12-17T15:15:00Z',
    pd.date_range(start="2005-05-10T18:00:00Z",end="2005-05-12T18:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2006-03-03T14:30:00Z',
    '2006-12-13T03:30:00Z',
    '2007-01-24T19:30:00Z',
    '2007-05-28T18:45:00Z',
    '2007-09-02T22:15:00Z',
    '2008-09-23T22:45:00Z',
    '2008-12-17T03:15:00Z',
    '2010-05-07T05:00:00Z',
    pd.date_range(start="2010-09-22T00:00:00Z",end="2010-09-23T16:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2011-05-26T18:30:00Z",end="2011-05-27T12:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2012-01-17T23:15:00Z',
    '2012-02-25T14:45:00Z',
    '2012-03-11T15:00:00Z',
    '2012-05-02T07:15:00Z',
    pd.date_range(start="2012-05-03T18:00:00Z",end="2012-05-03T22:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2013-12-31T00:45:00Z',
    '2017-02-08T16:30:00Z',
    pd.date_range(start="2017-09-02T13:15:00Z",end="2017-09-02T14:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-09-29T00:30:00Z',
    '2019-09-09T12:00:00Z',
    '2019-09-09T12:15:00Z',
    pd.date_range(start="2019-09-22T13:15:00Z",end="2019-09-22T13:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2020-02-11T15:00:00Z',
    '2020-04-25T22:15:00Z',
    '2020-06-21T10:45:00Z',
    '2020-06-24T09:00:00Z',
    '2020-09-06T07:30:00Z',
    '2020-09-06T21:15:00Z',
    pd.date_range(start="2020-09-19T00:45:00Z",end="2020-09-19T05:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2020-09-26T07:30:00Z",end="2020-09-26T10:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2020-10-07T04:45:00Z",end="2020-10-07T10:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2020-10-17T02:30:00Z',
    pd.date_range(start="2020-10-17T23:30:00Z",end="2020-10-18T12:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2020-10-19T02:30:00Z",end="2020-10-19T11:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
                             ],

         'Divisa' : [
    pd.date_range(start="1997-10-10T14:45:00Z",end="1997-10-10T16:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1998-02-13T18:45:00Z",end="1998-02-13T22:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1998-02-19T19:00:00Z",end="1998-02-19T22:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1998-11-18T15:45:00Z',
    pd.date_range(start="1999-02-11T19:15:00Z",end="1999-02-11T20:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1999-09-01T20:45:00Z",end="1999-09-08T20:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2000-06-27T09:30:00Z",end="2000-06-27T10:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2000-07-25T06:00:00Z",end="2000-07-25T06:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2000-07-25T10:30:00Z",end="2000-07-25T12:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-05-10T11:15:00Z",end="2001-05-10T12:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-05-12T14:00:00Z",end="2001-05-12T21:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-05-14T21:00:00Z",end="2001-05-15T02:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-05-28T23:00:00Z",end="2001-05-29T01:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-05-29T16:15:00Z",end="2001-05-29T18:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-06-14T12:30:00Z",end="2001-06-14T13:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-06-15T13:45:00Z",end="2001-06-15T14:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-06-18T02:15:00Z",end="2001-06-18T03:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2001-12-12T00:00:00Z',
    pd.date_range(start="2002-03-17T14:30:00Z",end="2002-03-17T14:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2002-03-17T17:00:00Z",end="2002-03-17T19:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2002-07-18T19:45:00Z',
    pd.date_range(start="2003-04-18T00:00:00Z",end="2003-04-18T02:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2003-12-03T10:45:00Z',
    '2004-01-25T22:30:00Z',
    '2004-04-12T08:30:00Z',
    '2004-05-08T06:45:00Z',
    '2005-06-08T13:45:00Z',
    pd.date_range(start="2006-07-26T16:00:00Z",end="2006-07-26T16:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2007-09-05T18:45:00Z',
    pd.date_range(start="2007-10-23T18:15:00Z",end="2007-10-24T10:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2007-10-25T11:00:00Z",end="2007-10-25T13:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-03-13T10:30:00Z',
    '2008-04-04T10:30:00Z',
    '2008-06-12T17:45:00Z',
    '2008-12-03T05:45:00Z',
    '2009-03-04T06:30:00Z',
    '2009-03-06T23:00:00Z',
    '2009-07-23T17:15:00Z',
    '2009-12-28T03:30:00Z',
    '2009-12-31T07:30:00Z',
    '2010-01-08T15:00:00Z',
    '2010-11-05T11:45:00Z',
    '2011-04-07T17:00:00Z',
    '2011-07-31T20:00:00Z',
    '2011-10-03T16:45:00Z',
    '2011-10-27T08:15:00Z',
    '2011-10-27T11:00:00Z',
    pd.date_range(start="2011-10-28T13:15:00Z",end="2011-10-28T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2011-11-27T23:00:00Z',
    '2013-02-26T14:15:00Z',
    '2013-02-28T01:30:00Z',
    pd.date_range(start="2013-02-28T10:15:00Z",end="2013-02-28T10:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2013-02-28T12:30:00Z',
    pd.date_range(start="2017-03-24T20:45:00Z",end="2017-03-24T21:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-03-25T21:30:00Z',
    pd.date_range(start="2017-06-05T01:30:00Z",end="2017-06-05T06:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2017-08-13T16:00:00Z",end="2017-08-13T17:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2018-01-12T11:30:00Z',
    pd.date_range(start="2018-01-12T12:45:00Z",end="2018-01-12T13:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2018-01-25T11:30:00Z',
    '2018-01-25T14:15:00Z',
    pd.date_range(start="2018-01-26T14:45:00Z",end="2018-01-26T21:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2018-01-29T13:15:00Z",end="2018-01-29T14:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2018-01-31T19:45:00Z",end="2018-01-31T20:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2018-02-08T15:30:00Z",end="2018-02-08T16:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2018-02-20T15:00:00Z",end="2018-02-20T17:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2018-02-23T14:00:00Z',
    '2018-02-23T18:30:00Z',
    '2018-02-23T19:00:00Z',
    pd.date_range(start="2018-02-26T13:00:00Z",end="2018-02-26T13:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2018-02-28T22:30:00Z',
    pd.date_range(start="2018-03-02T15:45:00Z",end="2018-03-02T21:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2019-05-02T11:15:00Z",end="2019-05-02T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2019-05-02T16:00:00Z",end="2019-05-02T16:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2019-05-04T16:45:00Z",end="2019-05-04T22:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-05-25T11:30:00Z',
    pd.date_range(start="2019-05-30T07:15:00Z",end="2019-05-30T07:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-05-30T08:45:00Z',
    pd.date_range(start="2019-05-30T22:45:00Z",end="2019-05-31T11:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-12-05T15:00:00Z',
    '2020-03-14T08:15:00Z',
    '2020-07-13T14:15:00Z',
    '2020-07-20T15:30:00Z',
    pd.date_range(start="2020-07-20T16:15:00Z",end="2020-07-20T16:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2020-07-20T17:15:00Z',
    '2020-07-20T17:45:00Z',
    pd.date_range(start="2020-07-22T17:45:00Z",end="2020-07-22T18:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2020-07-23T01:00:00Z",end="2020-07-23T01:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2020-09-05T01:45:00Z',
                     ],

         'Fluviopolis' : [
    '1997-06-07T10:45:00Z',
    '1997-06-13T13:15:00Z',
    '1997-06-13T14:30:00Z',
    pd.date_range(start="1997-07-16T19:45:00Z",end="1997-07-16T21:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1997-08-01T16:45:00Z",end="1997-08-01T19:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1997-10-10T12:00:00Z",end="1997-10-10T14:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1997-11-04T21:30:00Z',
    pd.date_range(start="1997-11-12T23:45:00Z",end="1997-11-13T02:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1999-03-26T16:00:00Z",end="1999-03-26T18:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1999-07-12T18:15:00Z',
    '1999-09-04T11:15:00Z',
    '1999-09-27T10:30:00Z',
    '1999-10-17T15:15:00Z',
    '1999-10-18T19:30:00Z',
    pd.date_range(start="1999-11-13T12:15:00Z",end="1999-11-13T12:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1999-11-17T17:45:00Z",end="1999-11-17T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2000-01-19T01:00:00Z',
    '2000-10-28T20:15:00Z',
    '2000-12-16T17:00:00Z',
    pd.date_range(start="2001-09-03T18:00:00Z",end="2001-09-03T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2001-10-01T23:45:00Z',
    '2001-12-02T23:30:00Z',
    '2002-01-22T12:00:00Z',
    pd.date_range(start="2003-02-27T12:15:00Z",end="2003-02-27T15:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2003-03-03T20:15:00Z",end="2003-03-03T23:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2003-03-10T08:15:00Z",end="2003-03-10T11:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2003-03-12T20:00:00Z",end="2003-03-12T23:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2003-03-18T04:00:00Z",end="2003-03-18T11:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2003-05-17T17:00:00Z",end="2003-05-17T19:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2003-05-30T19:30:00Z',
    '2003-09-23T21:00:00Z',
    '2004-02-07T10:15:00Z',
    '2004-04-09T03:15:00Z',
    '2004-04-23T21:15:00Z',
    '2004-04-28T19:15:00Z',
    '2004-06-24T01:45:00Z',
    '2004-07-02T06:45:00Z',
    '2004-08-21T18:45:00Z',
    '2004-09-03T10:00:00Z',
    '2004-10-04T18:45:00Z',
    '2004-10-14T08:00:00Z',
    '2004-10-19T04:45:00Z',
    '2004-11-06T16:00:00Z',
    '2004-12-04T07:45:00Z',
    '2005-02-10T06:45:00Z',
    '2005-05-17T04:30:00Z',
    '2005-07-07T18:30:00Z',
    '2005-10-05T21:00:00Z',
    '2006-07-24T04:30:00Z',
    '2006-08-31T10:30:00Z',
    '2006-09-10T18:00:00Z',
    '2007-09-13T03:00:00Z',
    '2007-11-05T08:00:00Z',
    '2008-06-19T15:30:00Z',
    '2008-07-03T23:15:00Z',
    '2008-07-05T22:45:00Z',
    '2008-07-31T14:00:00Z',
    pd.date_range(start="2008-08-13T12:30:00Z",end="2008-08-15T13:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-09-07T22:30:00Z',
    '2008-12-03T19:15:00Z',
    '2008-12-07T04:00:00Z',
    '2009-06-08T05:45:00Z',
    '2009-07-12T04:45:00Z',
    '2009-07-22T08:00:00Z',
    '2009-09-09T15:00:00Z',
    '2009-09-11T21:00:00Z',
    '2009-09-22T15:30:00Z',
    '2009-09-24T11:30:00Z',
    '2009-10-22T00:00:00Z',
    '2009-12-06T18:00:00Z',
    '2010-01-01T18:30:00Z',
    '2010-01-28T13:15:00Z',
    '2010-01-29T21:30:00Z',
    '2010-03-30T03:15:00Z',
    '2010-03-31T17:15:00Z',
    '2010-04-26T18:15:00Z',
    '2010-04-26T19:15:00Z',
    '2010-05-20T10:00:00Z',
    '2010-05-22T13:15:00Z',
    '2010-05-24T04:30:00Z',
    '2010-06-01T18:15:00Z',
    '2010-07-16T16:00:00Z',
    '2010-09-11T10:45:00Z',
    '2010-09-21T17:30:00Z',
    '2010-10-05T16:00:00Z',
    '2010-10-12T08:00:00Z',
    '2010-11-25T11:00:00Z',
    pd.date_range(start="2010-11-25T12:30:00Z",end="2010-11-29T06:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2011-10-01T00:15:00Z',
    '2012-02-17T22:45:00Z',
    pd.date_range(start="2012-07-29T12:45:00Z",end="2012-07-29T13:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2017-09-22T14:30:00Z",end="2017-09-22T16:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2018-03-07T15:00:00Z",end="2018-03-07T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
                          ],

         'Uniao_da_Vitoria' : [
    '1997-08-13T14:00:00Z',
    '1997-08-14T13:45:00Z',
    '1997-09-02T13:45:00Z',
    '1997-09-03T14:00:00Z',
    '1997-09-17T14:00:00Z',
    '1997-11-09T18:30:00Z',
    '1998-01-05T21:30:00Z',
    '1998-01-22T19:00:00Z',
    '1998-01-23T17:00:00Z',
    '1998-06-19T10:30:00Z',
    '1998-07-06T13:00:00Z',
    '1998-08-18T18:00:00Z',
    pd.date_range(start="1998-10-06T15:30:00Z",end="1998-10-07T15:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1998-10-30T11:15:00Z",end="1998-10-30T11:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1998-10-30T16:15:00Z",end="1998-10-30T16:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1999-03-02T16:30:00Z",end="1999-03-04T03:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1999-03-26T16:00:00Z",end="1999-03-26T18:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2000-01-30T04:00:00Z',
    '2000-06-13T19:15:00Z',
    '2000-06-13T20:00:00Z',
    '2001-02-11T03:45:00Z',
    '2001-04-04T03:15:00Z',
    pd.date_range(start="2002-04-04T16:30:00Z",end="2002-04-04T20:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2002-04-21T08:00:00Z',
    '2003-01-05T19:45:00Z',
    '2003-03-25T23:00:00Z',
    '2003-04-17T22:30:00Z',
    '2003-04-26T23:45:00Z',
    '2003-05-20T06:45:00Z',
    '2003-11-05T07:30:00Z',
    '2004-03-27T22:30:00Z',
    '2004-04-04T22:45:00Z',
    '2004-04-19T05:00:00Z',
    '2004-06-29T22:15:00Z',
    pd.date_range(start="2004-07-23T14:30:00Z",end="2004-07-23T19:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2004-08-01T04:45:00Z",end="2004-08-06T19:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2004-11-07T23:30:00Z',
    '2005-05-26T19:15:00Z',
    '2006-03-05T19:45:00Z',
    '2006-03-15T18:15:00Z',
    '2006-04-10T10:00:00Z',
    '2006-04-28T19:15:00Z',
    '2006-05-03T06:45:00Z',
    '2007-01-11T23:15:00Z',
    '2007-01-22T11:15:00Z',
    '2007-04-07T19:00:00Z',
    '2007-04-29T07:15:00Z',
    pd.date_range(start="2007-07-09T20:00:00Z",end="2007-07-09T20:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2007-12-16T14:45:00Z',
    '2008-04-29T15:00:00Z',
    '2008-08-04T11:45:00Z',
    '2008-09-24T03:15:00Z',
    '2009-02-21T15:15:00Z',
    '2009-02-21T16:00:00Z',
    '2009-02-21T18:15:00Z',
    '2009-02-21T19:00:00Z',
    '2009-02-21T20:15:00Z',
    '2009-02-21T22:00:00Z',
    pd.date_range(start="2009-02-22T14:30:00Z",end="2009-02-22T14:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2009-02-22T16:30:00Z',
    '2009-02-22T23:45:00Z',
    '2009-02-23T16:45:00Z',
    pd.date_range(start="2009-02-23T17:15:00Z",end="2009-02-23T17:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2009-02-23T19:00:00Z',
    '2009-02-23T19:45:00Z',
    '2009-02-24T03:45:00Z',
    pd.date_range(start="2009-02-24T13:00:00Z",end="2009-02-24T13:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2009-02-24T14:15:00Z',
    '2009-02-24T16:15:00Z',
    '2009-02-24T18:15:00Z',
    '2009-02-24T20:45:00Z',
    '2009-02-24T23:00:00Z',
    '2009-02-25T00:45:00Z',
    '2009-02-25T01:30:00Z',
    pd.date_range(start="2009-02-25T02:15:00Z",end="2009-02-25T02:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2009-02-25T03:15:00Z",end="2009-02-25T03:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2009-02-25T04:00:00Z",end="2009-02-25T04:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2009-02-25T06:45:00Z",end="2009-02-25T07:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2009-02-25T07:45:00Z',
    '2009-02-25T08:30:00Z',
    '2009-02-25T09:30:00Z',
    '2009-02-25T10:00:00Z',
    '2009-02-25T12:15:00Z',
    pd.date_range(start="2009-02-25T12:45:00Z",end="2009-02-25T13:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2009-02-25T13:30:00Z',
    '2009-02-25T15:45:00Z',
    '2009-04-17T02:45:00Z',
    '2009-05-27T09:30:00Z',
    pd.date_range(start="2009-06-08T16:00:00Z",end="2009-06-08T16:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2009-06-08T19:00:00Z",end="2009-06-08T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2011-03-31T00:00:00Z",end="2011-03-31T11:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2013-04-12T13:00:00Z",end="2013-04-12T13:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2013-07-13T12:30:00Z",end="2013-07-13T12:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2013-07-14T17:15:00Z",end="2013-07-14T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2015-02-21T15:45:00Z',
    pd.date_range(start="2015-03-12T23:30:00Z",end="2015-03-12T23:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2016-06-15T06:45:00Z',
    '2017-04-04T19:15:00Z',
    '2017-04-21T14:00:00Z',
    pd.date_range(start="2017-04-21T15:15:00Z",end="2017-04-21T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-04-22T12:15:00Z',
    pd.date_range(start="2017-04-22T12:45:00Z",end="2017-04-22T13:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-04-22T19:30:00Z',
    pd.date_range(start="2017-04-22T21:15:00Z",end="2017-04-22T21:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-04-23T06:45:00Z',
    '2017-04-23T09:15:00Z',
    '2017-04-23T20:45:00Z',
    '2017-04-24T06:00:00Z',
    '2017-04-24T07:15:00Z',
    '2017-04-24T11:00:00Z',
    '2017-04-24T13:45:00Z',
    '2017-04-24T17:30:00Z',
    '2017-04-24T19:15:00Z',
    '2017-04-24T21:15:00Z',
    '2017-04-25T05:15:00Z',
    '2017-04-25T05:45:00Z',
    pd.date_range(start="2017-04-25T09:00:00Z",end="2017-04-25T09:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2017-04-25T10:00:00Z",end="2017-04-25T10:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-04-25T12:00:00Z',
    '2017-04-25T18:00:00Z',
    '2017-04-25T19:00:00Z',
    '2017-04-25T21:00:00Z',
    '2017-04-25T22:00:00Z',
    '2017-04-26T01:45:00Z',
    pd.date_range(start="2017-04-26T04:15:00Z",end="2017-04-26T04:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2017-04-26T16:00:00Z",end="2017-04-26T16:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-05T15:15:00Z',
    '2017-05-05T16:15:00Z',
    '2017-05-05T20:00:00Z',
    '2017-05-06T13:00:00Z',
    '2017-05-06T14:45:00Z',
    pd.date_range(start="2017-05-06T16:15:00Z",end="2017-05-06T16:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-06T17:30:00Z',
    '2017-05-06T19:15:00Z',
    '2017-05-07T05:45:00Z',
    pd.date_range(start="2017-05-07T09:00:00Z",end="2017-05-07T09:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2017-05-07T15:45:00Z",end="2017-05-07T16:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-07T18:00:00Z',
    pd.date_range(start="2017-05-07T19:45:00Z",end="2017-05-07T20:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-08T10:00:00Z',
    '2017-05-08T14:30:00Z',
    '2017-05-08T17:15:00Z',
    '2017-05-09T14:30:00Z',
    '2017-05-09T15:15:00Z',
    '2017-05-09T15:45:00Z',
    pd.date_range(start="2017-05-09T21:45:00Z",end="2017-05-09T22:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-10T03:00:00Z',
    '2017-05-10T07:45:00Z',
    '2017-05-10T11:00:00Z',
    '2017-05-10T13:30:00Z',
    pd.date_range(start="2017-05-10T16:30:00Z",end="2017-05-10T16:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-10T18:15:00Z',
    pd.date_range(start="2017-05-10T18:45:00Z",end="2017-05-10T19:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-11T03:45:00Z',
    '2017-05-11T06:00:00Z',
    '2017-05-11T06:30:00Z',
    '2017-05-11T10:45:00Z',
    '2017-05-11T11:30:00Z',
    '2017-05-11T14:15:00Z',
    '2017-05-11T15:45:00Z',
    '2017-05-11T18:15:00Z',
    '2017-05-11T21:15:00Z',
    '2017-05-11T22:45:00Z',
    pd.date_range(start="2017-05-12T00:30:00Z",end="2017-05-12T01:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-12T04:00:00Z',
    '2017-05-12T04:45:00Z',
    '2017-05-12T14:00:00Z',
    '2017-05-12T15:15:00Z',
    '2017-05-12T15:45:00Z',
    pd.date_range(start="2017-05-13T15:00:00Z",end="2017-05-13T15:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-13T16:00:00Z',
    '2017-05-13T17:00:00Z',
    '2017-05-13T18:00:00Z',
    pd.date_range(start="2017-05-13T18:45:00Z",end="2017-05-13T19:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-13T23:30:00Z',
    '2017-05-14T12:00:00Z',
    '2017-05-14T22:45:00Z',
    '2017-05-15T19:15:00Z',
    pd.date_range(start="2017-05-16T07:45:00Z",end="2017-05-16T08:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2017-05-17T11:00:00Z",end="2017-05-17T11:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-05-17T13:30:00Z',
    '2017-05-17T20:15:00Z',
    '2017-05-17T21:00:00Z',
    '2017-05-18T03:15:00Z',
    '2017-05-18T11:00:00Z',
    '2017-05-18T14:45:00Z',
    pd.date_range(start="2017-07-07T13:45:00Z",end="2017-07-07T14:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2017-08-13T14:45:00Z",end="2017-08-14T10:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2018-08-09T20:30:00Z",end="2018-08-09T21:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2018-08-09T22:15:00Z',
    '2018-08-09T23:15:00Z',
    '2018-08-10T00:15:00Z',
    '2018-08-10T01:15:00Z',
    '2018-08-10T00:15:00Z',
    '2018-08-10T01:15:00Z',
    '2018-08-10T02:15:00Z',
    '2018-08-10T03:15:00Z',
    '2018-08-10T04:15:00Z',
    '2018-08-10T05:15:00Z',
    '2018-08-10T06:15:00Z',
    '2018-08-10T07:15:00Z',
    pd.date_range(start="2018-08-10T07:45:00Z",end="2018-08-10T08:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2018-08-10T09:00:00Z',
    '2018-08-10T10:00:00Z',
    '2018-08-10T11:00:00Z',
    '2018-08-10T11:45:00Z',
    '2018-08-10T13:00:00Z',
    '2018-08-10T14:00:00Z',
    '2018-08-10T15:00:00Z',
    '2018-08-10T16:00:00Z',
    '2018-08-10T17:00:00Z',
    '2018-08-10T18:00:00Z',
    '2018-08-10T19:00:00Z',
    '2018-08-10T20:00:00Z',
    '2018-08-10T23:00:00Z',
    pd.date_range(start="2018-08-10T23:45:00Z",end="2018-08-11T11:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2018-08-12T00:30:00Z',
    '2019-06-22T01:15:00Z',
    '2019-06-28T22:15:00Z',
    '2019-06-28T23:15:00Z',
    '2019-07-03T08:30:00Z',
    '2019-12-16T20:30:00Z',
    '2019-12-16T21:30:00Z',
    '2019-12-17T16:15:00Z',
    '2019-12-17T18:00:00Z',
    '2019-12-17T18:30:00Z',
    pd.date_range(start="2019-12-17T23:45:00Z",end="2019-12-18T00:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-12-18T06:00:00Z',
    pd.date_range(start="2019-12-18T06:45:00Z",end="2019-12-18T07:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-12-18T07:45:00Z',
    '2019-12-18T08:15:00Z',
    '2019-12-18T08:45:00Z',
    pd.date_range(start="2019-12-18T10:00:00Z",end="2019-12-18T10:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-12-18T12:15:00Z',
    '2019-12-18T19:00:00Z',
    '2019-12-19T01:00:00Z',
    '2019-12-19T01:30:00Z',
    '2019-12-19T02:15:00Z',
    pd.date_range(start="2019-12-19T03:30:00Z",end="2019-12-19T03:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2019-12-19T04:45:00Z',
    '2019-12-19T06:15:00Z',
    '2019-12-19T06:45:00Z',
    pd.date_range(start="2019-12-19T11:30:00Z",end="2019-12-19T14:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
                               ],

         'Madereira_Gavazzoni' : [
    pd.date_range(start="1997-06-14T08:45:00Z",end="1997-06-18T17:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1997-06-21T12:00:00Z',
    '1997-09-19T16:30:00Z',
    pd.date_range(start="1997-10-09T17:45:00Z",end="1997-11-06T20:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1997-11-12T01:00:00Z',
    pd.date_range(start="1998-11-19T16:00:00Z",end="1998-11-20T11:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1999-03-26T16:00:00Z",end="1999-03-26T18:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1999-04-03T02:15:00Z',
    pd.date_range(start="1999-12-27T20:00:00Z",end="1999-12-27T20:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2000-01-01T00:45:00Z',
    '2000-11-28T00:00:00Z',
    '2001-02-22T23:45:00Z',
    '2001-11-21T22:00:00Z',
    '2002-06-06T07:45:00Z',
    '2002-10-15T07:45:00Z',
    '2003-06-15T14:45:00Z',
    '2003-10-23T23:30:00Z',
    '2003-12-09T06:45:00Z',
    '2003-12-19T15:15:00Z',
    '2004-06-16T09:30:00Z',
    '2004-07-17T19:15:00Z',
    '2004-09-07T09:15:00Z',
    '2004-11-03T06:15:00Z',
    pd.date_range(start="2004-11-23T17:15:00Z",end="2004-11-25T02:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2004-11-28T01:45:00Z",end="2004-11-28T02:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2004-11-28T07:30:00Z",end="2004-11-28T08:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2004-11-28T16:30:00Z",end="2004-11-28T17:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2004-11-28T20:00:00Z",end="2004-11-28T21:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2004-12-02T04:45:00Z",end="2004-12-02T05:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2004-12-03T00:30:00Z",end="2004-12-03T13:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2004-12-08T14:30:00Z',
    pd.date_range(start="2005-02-12T08:15:00Z",end="2005-02-16T19:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2005-03-27T03:30:00Z',
    pd.date_range(start="2005-06-30T18:00:00Z",end="2005-06-30T18:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2005-07-03T10:45:00Z',
    '2005-07-28T18:45:00Z',
    '2006-05-27T19:15:00Z',
    '2006-06-28T07:30:00Z',
    '2006-07-06T14:30:00Z',
    '2006-07-13T13:30:00Z',
    '2006-07-13T15:15:00Z',
    '2006-07-16T23:30:00Z',
    '2007-01-03T00:15:00Z',
    '2007-09-22T19:00:00Z',
    '2007-11-01T04:30:00Z',
    '2007-12-20T00:45:00Z',
    '2008-02-13T15:30:00Z',
    pd.date_range(start="2008-10-21T02:45:00Z",end="2008-10-21T03:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2009-03-27T01:00:00Z',
    '2009-03-28T05:00:00Z',
    '2009-04-13T11:15:00Z',
    '2009-07-23T12:15:00Z',
    '2009-08-08T21:30:00Z',
    '2009-08-13T13:30:00Z',
    '2009-09-08T10:00:00Z',
    '2009-09-25T03:15:00Z',
    '2009-10-04T01:30:00Z',
    '2009-10-12T22:15:00Z',
    '2009-10-13T08:00:00Z',
    '2009-10-15T05:00:00Z',
    '2009-10-26T21:30:00Z',
    '2009-11-03T11:00:00Z',
    '2009-12-16T13:30:00Z',
    '2010-05-06T13:30:00Z',
    '2010-05-06T22:30:00Z',
    '2010-07-16T03:00:00Z',
    '2010-09-29T15:15:00Z',
    '2010-10-28T16:45:00Z',
    '2013-04-14T04:45:00Z',
    pd.date_range(start="2013-10-20T23:00:00Z",end="2013-10-27T00:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2020-09-01T18:00:00Z",end="2020-09-01T20:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
                                  ],

         'Jangada' : [
    pd.date_range(start="1997-07-17T10:15:00Z",end="1997-07-19T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1997-08-20T11:30:00Z",end="1997-08-25T12:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1997-09-02T05:00:00Z",end="1997-09-12T14:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1997-09-14T00:45:00Z',
    pd.date_range(start="1997-10-10T12:15:00Z",end="1997-10-10T13:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1997-11-07T07:15:00Z',
    pd.date_range(start="1998-10-30T10:15:00Z",end="1998-11-02T03:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1999-03-19T14:45:00Z',
    '1999-03-29T18:30:00Z',
    '2000-12-06T01:00:00Z',
    '2001-01-06T22:45:00Z',
    '2001-01-22T02:15:00Z',
    '2003-06-16T10:45:00Z',
    '2003-07-31T09:30:00Z',
    '2003-12-01T03:00:00Z',
    '2004-03-20T19:15:00Z',
    '2005-05-31T11:15:00Z',
    '2005-06-25T19:15:00Z',
    pd.date_range(start="2005-09-14T14:00:00Z",end="2005-09-14T15:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2005-11-03T18:45:00Z',
    '2006-09-23T23:00:00Z',
    '2007-02-23T22:45:00Z',
    pd.date_range(start="2007-05-30T12:00:00Z",end="2007-05-31T15:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2007-12-22T22:45:00Z',
    '2008-02-07T03:00:00Z',
    '2008-04-06T05:30:00Z',
    '2008-04-22T13:30:00Z',
    pd.date_range(start="2008-05-16T14:45:00Z",end="2008-05-20T17:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-06-27T23:30:00Z',
    '2009-01-11T15:45:00Z',
    '2009-01-12T14:45:00Z',
    '2009-04-27T17:45:00Z',
    '2009-07-31T17:30:00Z',
    pd.date_range(start="2009-10-01T14:30:00Z",end="2009-10-01T14:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2009-11-05T09:30:00Z',
    pd.date_range(start="2009-11-24T21:45:00Z",end="2009-11-25T00:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2009-11-25T12:00:00Z",end="2009-11-25T12:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2009-11-25T15:15:00Z",end="2009-11-25T15:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2009-11-25T17:00:00Z',
    '2009-11-26T14:30:00Z',
    pd.date_range(start="2009-12-01T19:15:00Z",end="2009-12-02T10:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2010-08-07T13:15:00Z',
    pd.date_range(start="2010-09-23T05:00:00Z",end="2010-09-27T20:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2010-09-28T13:15:00Z',
    '2010-10-10T09:00:00Z',
    pd.date_range(start="2010-11-17T01:15:00Z",end="2010-11-17T01:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2011-01-20T18:30:00Z",end="2011-01-21T10:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2011-07-02T03:00:00Z',
    pd.date_range(start="2011-07-11T20:15:00Z",end="2011-07-12T15:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2011-07-25T14:15:00Z',
    '2011-07-25T21:45:00Z',
    '2011-07-26T13:15:00Z',
    '2011-07-26T21:15:00Z',
    '2011-07-27T14:30:00Z',
    '2011-07-28T13:45:00Z',
    '2011-07-28T20:15:00Z',
    '2011-07-28T21:15:00Z',
    '2011-07-30T19:30:00Z',
    '2011-08-16T20:00:00Z',
    '2011-08-17T14:00:00Z',
    '2011-08-17T21:15:00Z',
    '2011-08-18T13:45:00Z',
    '2011-08-29T13:45:00Z',
    pd.date_range(start="2011-09-12T21:00:00Z",end="2011-09-13T14:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2011-11-09T00:00:00Z',
    '2011-11-21T00:30:00Z',
    pd.date_range(start="2012-04-21T07:00:00Z",end="2012-04-24T17:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2017-06-21T16:30:00Z',
    '2017-07-28T12:15:00Z',
    pd.date_range(start="2018-01-08T18:45:00Z",end="2018-01-08T19:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2020-08-23T11:30:00Z',
    '2020-12-04T01:00:00Z',
    '2020-12-04T08:30:00Z',
    '2020-12-07T04:15:00Z',
    '2020-12-14T21:45:00Z',
                      ],

         'Solais_Novo' : [
    pd.date_range(start="1997-05-01T10:00:00Z",end="1997-11-07T20:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1998-03-28T20:00:00Z",end="1998-04-09T10:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="1998-08-22T20:00:00Z",end="1999-03-27T23:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2000-02-27T20:00:00Z',
    '2000-02-28T10:00:00Z',
    '2000-02-28T20:00:00Z',
    pd.date_range(start="2000-08-26T20:00:00Z",end="2000-10-12T10:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2001-03-07T01:45:00Z',
    '2002-08-27T16:45:00Z',
    '2002-08-27T18:15:00Z',
    pd.date_range(start="2002-09-05T18:15:00Z",end="2002-09-06T02:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2003-05-16T15:00:00Z',
    '2004-03-01T08:45:00Z',
    '2004-06-28T14:30:00Z',
    '2004-08-11T18:45:00Z',
    '2004-08-22T08:15:00Z',
    '2005-05-07T05:45:00Z',
    '2005-05-19T22:00:00Z',
    '2005-10-08T21:30:00Z',
    '2005-10-26T23:30:00Z',
    '2006-02-15T16:00:00Z',
    pd.date_range(start="2007-05-09T16:00:00Z",end="2007-05-16T01:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2007-06-03T21:30:00Z",end="2007-06-28T17:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-11-05T05:45:00Z',
    pd.date_range(start="2008-12-02T14:00:00Z",end="2008-12-02T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2009-01-08T18:00:00Z',
    '2009-01-11T03:30:00Z',
    '2009-01-21T04:00:00Z',
    '2009-02-03T14:30:00Z',
    '2009-02-15T08:45:00Z',
    '2009-02-24T06:15:00Z',
    '2009-03-07T04:15:00Z',
    '2009-03-08T14:45:00Z',
    '2009-03-09T11:15:00Z',
    '2009-03-17T21:30:00Z',
    '2009-04-15T20:30:00Z',
    '2009-04-22T15:30:00Z',
    '2009-04-27T01:15:00Z',
    pd.date_range(start="2009-06-10T18:15:00Z",end="2009-06-10T19:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2011-01-05T09:30:00Z',
    '2011-01-09T02:00:00Z',
    '2011-05-03T18:45:00Z',
    pd.date_range(start="2012-01-01T17:00:00Z",end="2012-01-01T17:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2012-04-05T17:30:00Z',
    pd.date_range(start="2012-04-11T16:00:00Z",end="2012-04-11T16:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2013-02-07T02:45:00Z',
    '2013-10-18T23:45:00Z',
    pd.date_range(start="2014-06-08T17:45:00Z",end="2014-06-10T08:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2014-06-24T15:30:00Z",end="2014-06-24T16:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2014-12-28T22:15:00Z',
    '2019-08-13T14:45:00Z',
    pd.date_range(start="2019-08-30T19:30:00Z",end="2019-08-30T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2020-02-24T18:30:00Z",end="2020-02-24T20:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2020-05-13T15:30:00Z',
    pd.date_range(start="2020-11-04T05:15:00Z",end="2020-11-04T11:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2020-12-13T18:45:00Z',
    pd.date_range(start="2020-12-15T19:30:00Z",end="2020-12-15T19:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
                          ],

         'Porto_Santo_Antonio' : [
    '1997-06-18T21:15:00Z',
    pd.date_range(start="1997-06-27T19:15:00Z",end="1997-06-27T20:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1997-07-27T11:15:00Z',
    '1997-10-10T16:30:00Z',
    '1999-04-05T03:15:00Z',
    pd.date_range(start="2000-03-30T20:30:00Z",end="2000-03-31T11:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2000-06-02T15:00:00Z',
    '2000-06-13T22:15:00Z',
    '2000-06-13T22:45:00Z',
    '2000-07-11T02:00:00Z',
    '2000-07-12T12:30:00Z',
    pd.date_range(start="2000-10-11T04:45:00Z",end="2000-10-11T15:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2000-11-08T20:45:00Z',
    '2001-03-03T02:15:00Z',
    pd.date_range(start="2001-03-03T09:30:00Z",end="2001-03-03T11:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2001-03-03T16:00:00Z',
    pd.date_range(start="2001-10-16T18:00:00Z",end="2001-10-16T18:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2002-04-29T17:15:00Z',
    '2002-10-26T19:15:00Z',
    '2004-03-18T00:00:00Z',
    '2004-07-20T12:00:00Z',
    pd.date_range(start="2004-08-20T12:15:00Z",end="2004-08-20T17:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2004-09-20T12:15:00Z',
    '2005-05-09T13:00:00Z',
    '2005-08-11T21:30:00Z',
    '2006-11-22T17:30:00Z',
    '2007-05-06T09:15:00Z',
    '2007-05-25T21:00:00Z',
    pd.date_range(start="2007-11-08T18:00:00Z",end="2007-11-08T18:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-02-18T09:15:00Z',
    pd.date_range(start="2008-02-27T21:00:00Z",end="2008-02-29T08:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-03-18T01:00:00Z',
    pd.date_range(start="2008-09-16T13:45:00Z",end="2008-09-16T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2008-10-22T21:15:00Z',
    '2008-11-25T00:45:00Z',
    '2008-12-07T23:00:00Z',
    '2009-01-25T09:30:00Z',
    '2009-02-12T21:15:00Z',
    '2009-05-24T20:30:00Z',
    '2009-07-03T18:15:00Z',
    pd.date_range(start="2009-07-06T20:00:00Z",end="2009-07-07T10:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2009-09-18T10:30:00Z',
    pd.date_range(start="2010-03-18T18:00:00Z",end="2010-03-21T04:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2010-03-24T01:00:00Z',
    '2010-03-25T01:00:00Z',
    '2010-03-25T11:00:00Z',
    '2010-03-30T03:15:00Z',
    '2010-04-02T08:30:00Z',
    '2010-04-02T12:30:00Z',
    '2010-04-06T02:00:00Z',
    '2010-04-08T00:00:00Z',
    '2010-04-24T04:15:00Z',
    '2010-04-24T06:15:00Z',
    '2010-05-05T09:15:00Z',
    '2010-05-06T22:15:00Z',
    '2010-07-15T18:45:00Z',
    pd.date_range(start="2010-09-01T15:15:00Z",end="2010-09-02T04:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2010-09-14T11:15:00Z',
    '2010-11-25T16:00:00Z',
    '2012-03-01T01:00:00Z',
    '2013-12-18T23:45:00Z',
    pd.date_range(start="2014-06-26T18:45:00Z",end="2014-06-28T10:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2014-09-11T14:15:00Z",end="2014-09-11T15:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2015-03-05T02:00:00Z',
         ],

         'Aguas_do_Vere' : [
    pd.date_range(start="1997-06-24T20:45:00Z",end="1997-06-25T00:30:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '1997-07-11T18:00:00Z',
    '1997-08-15T18:00:00Z',
    '1997-08-18T10:45:00Z',
    '1997-09-02T20:00:00Z',
    '1999-04-03T03:15:00Z',
    '2000-02-17T02:45:00Z',
    pd.date_range(start="2000-10-05T13:45:00Z",end="2000-10-05T14:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-05-29T12:00:00Z",end="2001-05-31T21:00:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    pd.date_range(start="2001-06-01T14:30:00Z",end="2001-06-01T14:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2001-12-01T00:15:00Z',
    '2002-04-11T05:15:00Z',
    '2002-08-23T16:15:00Z',
    pd.date_range(start="2002-08-28T13:30:00Z",end="2002-08-28T13:45:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2002-09-07T20:45:00Z',
    pd.date_range(start="2003-05-16T14:00:00Z",end="2003-05-16T14:15:00Z",
                  freq='15min').format(formatter=lambda x: x.strftime(
                      '%Y-%m-%dT%H:%M:%SZ')),
    '2003-05-24T08:00:00Z',
    '2004-01-05T23:30:00Z',
    '2004-03-18T12:00:00Z',
    '2004-04-05T23:45:00Z',
    '2004-12-04T12:30:00Z',
    '2005-07-25T00:15:00Z',
    '2005-10-12T07:30:00Z',
    '2005-11-25T00:00:00Z',
    
                            ],

         'Porto_Capanema' : [
                             ],

         'Hotel_Cataratas' : [
                              ]

                  }

# json.dump(erros, open(
#     '/discolocal/bruno/Observado/Pre_Consistencia/erros_grosseiros_bruto.txt',
#     'w'),indent = 2)
json.dump(erros, open(
    '../dados/erros_grosseiros_bruto.txt',
    'w'),indent = 2)
